Move Jetty code to separate module
